+++ res/os/hard/Keyboard.js
@@ -1,0 +1,23 @@
+import { globalShortcut } from "electron";
+import { shortcutInit } from "../controllers/ShortcutController";
+import { templateInit } from "../controllers/TemplateController";
+import { clipboardInit } from "../controllers/ClipboardController";
+
+//キーボードショートカット設定
+export function keyboardSetting(DbSet) {
+  const ClipboardOpenButton = "CommandOrControl+Shift+Z";
+  const ShortcutOpenButton = "CommandOrControl+Shift+X";
+  const TemplateOpenButton = "CommandOrControl+Shift+C";
+
+  globalShortcut.register(ClipboardOpenButton, () => {
+    clipboardInit(DbSet.InMemoryDb, DbSet.ClipboardDb);
+  });
+
+  globalShortcut.register(ShortcutOpenButton, () => {
+    shortcutInit(DbSet.InMemoryDb, DbSet.ShortcutDb);
+  });
+
+  globalShortcut.register(TemplateOpenButton, () => {
+    templateInit(DbSet.InMemoryDb, DbSet.TemplateDb);
+  });
+}
