+++ res/os/hard/Tray.js
@@ -1,0 +1,21 @@
+import { app, Menu, Tray } from "electron";
+
+//タスクトレイ設定
+export function trayInit() {
+  const tray = new Tray(__dirname + "/icon/icon.png");
+  var contextMenu = Menu.buildFromTemplate([
+    {
+      label: "設定",
+      click: () => {
+        createWindow();
+      },
+    },
+    {
+      label: "終了",
+      click: function () {
+        app.quit();
+      },
+    },
+  ]);
+  tray.setContextMenu(contextMenu);
+}
