+++ res/os/hard/Window.js
